pyproject.toml
[project]
name = "telegram-fastapi-app"
version = "0.1.0"
description = "FastAPI app for Telegram bot webhooks"
authors = [
    {name = "User"}
]
readme = "README.md"
requires-python = "^3.9"
dependencies = [
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.36.0,<0.37.0)",
    "python-telegram-bot (>=22.4,<23.0)",
    "pydantic (>=2.11.9,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
